+++ modules/canvaskit/tests/core.spec.js
@@ -165,24 +165,30 @@ describe('Core canvas behavior', () => {
         expect(aImg.width()).toEqual(320);
         expect(aImg.height()).toEqual(240);
         expect(aImg.getFrameCount()).toEqual(60);
-        // TODO(kjlubick): deprecate drawAnimatedImage and have it just snap off
-        // an SkImage at the desired frame.
 
-        canvas.drawAnimatedImage(aImg, 0, 0);
+        let img = aImg.getCurrentFrame();
+        canvas.drawImage(img, 0, 0, null);
+        img.delete();
 
         let c = aImg.decodeNextFrame();
         expect(c).not.toEqual(-1);
-        canvas.drawAnimatedImage(aImg, 300, 0);
+        img = aImg.getCurrentFrame();
+        canvas.drawImage(img, 300, 0, null);
+        img.delete();
         for(let i = 0; i < 10; i++) {
             c = aImg.decodeNextFrame();
             expect(c).not.toEqual(-1);
         }
-        canvas.drawAnimatedImage(aImg, 0, 300);
+        img = aImg.getCurrentFrame();
+        canvas.drawImage(img, 0, 300, null);
+        img.delete();
         for(let i = 0; i < 10; i++) {
             c = aImg.decodeNextFrame();
             expect(c).not.toEqual(-1);
         }
-        canvas.drawAnimatedImage(aImg, 300, 300);
+        img = aImg.getCurrentFrame();
+        canvas.drawImage(img, 300, 300, null);
+        img.delete();
 
         aImg.delete();
     }, '/assets/flightAnim.gif');